move GraphManager methods to QueryHandler
